fix: single resource for roles and permissions
