+++ internal/eva/role.go
@@ -10,10 +10,12 @@ import (
 )
 
 const (
-	createRolePath = "/api/core/management/CreateRole"
-	getRolePath    = "/api/core/management/GetRole"
-	updateRolePath = "/api/core/management/UpdateRole"
-	deleteRolePath = "/api/core/management/DeleteRole"
+	createRolePath                    = "/api/core/management/CreateRole"
+	getRolePath                       = "/api/core/management/GetRole"
+	updateRolePath                    = "/api/core/management/UpdateRole"
+	deleteRolePath                    = "/api/core/management/DeleteRole"
+	attachFunctionalitiesToRolePath   = "/api/core/management/AttachFunctionalitiesToRole"
+	detachFunctionalitiesFromRolePath = "/api/core/management/DetachFunctionalitiesFromRole"
 )
 
 type CreateRoleRequest struct {
@@ -162,6 +164,75 @@ func (c *Client) DeleteRole(ctx context.Context, req DeleteRoleRequest) (*Delete
 	if err := json.Unmarshal([]byte(resp.Body()), &jsonResp); err != nil {
 
 		return nil, errors.New(fmt.Sprintf("Response could not be parsed. Received: %s", resp.String()))
+	}
+
+	return &jsonResp, nil
+}
+
+type AttachFunctionalitiesToRoleRequest struct {
+	RoleID                int64                     `json:"RoleID"`
+	ScopedFunctionalities []RoleScopedFunctionality `json:"ScopedFunctionalities,omitempty"`
+}
+
+type AttachFunctionalitiesToRoleResponse struct {
+}
+
+func (c *Client) AttachFunctionalitiesToRole(ctx context.Context, req AttachFunctionalitiesToRoleRequest) (*AttachFunctionalitiesToRoleResponse, error) {
+	resp, err := c.restClient.R().
+		SetBody(req).
+		Post(attachFunctionalitiesToRolePath)
+
+	if err != nil {
+		tflog.Error(ctx, "An network error ocurred.", err)
+
+		return nil, err
+	}
+
+	if resp.StatusCode() != 200 {
+		tflog.Info(ctx, "Request failed", "Status code", resp.StatusCode(), "body", resp.String())
+
+		return nil, errors.New(fmt.Sprintf("Request failed with error: %s", resp.String()))
+	}
+
+	var jsonResp AttachFunctionalitiesToRoleResponse
+	if err := json.Unmarshal([]byte(resp.Body()), &jsonResp); err != nil {
+		return nil, errors.New(fmt.Sprintf("Response could not be parsed. Received: %s", resp.String()))
+	}
+
+	return &jsonResp, nil
+}
+
+type DetachFunctionalitiesFromRoleRequest struct {
+	RoleID                int64                     `json:"RoleID"`
+	ScopedFunctionalities []RoleScopedFunctionality `json:"ScopedFunctionalities,omitempty"`
+}
+
+type DetachFunctionalitiesFromRoleResponse struct {
+}
+
+func (c *Client) DetachFunctionalitiesFromRole(ctx context.Context, req DetachFunctionalitiesFromRoleRequest) (*DetachFunctionalitiesFromRoleResponse, error) {
+	resp, err := c.restClient.R().
+		SetBody(req).
+		Post(detachFunctionalitiesFromRolePath)
+
+	if err != nil {
+		tflog.Error(ctx, "An network error ocurred.", err)
+
+		return nil, err
+	}
+
+	if resp.StatusCode() != 200 {
+		tflog.Info(ctx, "Request failed", "Status code", resp.StatusCode(), "body", resp.String())
+
+		return nil, errors.New("Request failed.")
+	}
+
+	tflog.Debug(ctx, "Request info", "Status code", resp.StatusCode(), "body", resp.String())
+
+	var jsonResp DetachFunctionalitiesFromRoleResponse
+	if err := json.Unmarshal([]byte(resp.Body()), &jsonResp); err != nil {
+
+		return nil, errors.New(fmt.Sprintf("Response could not be parsed. Error: %s \n Received: %s", err, resp.String()))
 	}
 
 	return &jsonResp, nil
