+++ internal/provider/resource_eva_role.go
@@ -43,6 +43,30 @@ func (t roleType) GetSchema(ctx context.Context) (tfsdk.Schema, diag.Diagnostics
 				Optional:            true,
 				Type:                types.StringType,
 			},
+			"scoped_functionalities": {
+				MarkdownDescription: "list of scoped functionalities to be attached",
+				Required:            true,
+				Attributes: tfsdk.ListNestedAttributes(
+					map[string]tfsdk.Attribute{
+						"functionality": {
+							MarkdownDescription: "functionality identifier",
+							Required:            true,
+							Type:                types.StringType,
+						},
+						"scope": {
+							MarkdownDescription: "functionality scope",
+							Required:            true,
+							Type:                types.Int64Type,
+						},
+						"requires_elevation": {
+							MarkdownDescription: "whether functionality requires elevation or not",
+							Required:            true,
+							Type:                types.BoolType,
+						},
+					},
+					tfsdk.ListNestedAttributesOptions{},
+				),
+			},
 		},
 	}, nil
 }
@@ -59,15 +83,46 @@ type roleResource struct {
 	provider provider
 }
 
-type inputRoleData struct {
-	ID       types.Int64  `tfsdk:"id"`
-	Name     types.String `tfsdk:"name"`
-	UserType types.Int64  `tfsdk:"user_type"`
-	Code     types.String `tfsdk:"code"`
+type inputRole struct {
+	ID                    types.Int64                    `tfsdk:"id"`
+	Name                  types.String                   `tfsdk:"name"`
+	UserType              types.Int64                    `tfsdk:"user_type"`
+	Code                  types.String                   `tfsdk:"code"`
+	ScopedFunctionalities []inputRoleScopedFunctionality `tfsdk:"scoped_functionalities"`
+}
+
+type inputRoleScopedFunctionality struct {
+	Functionality     types.String `tfsdk:"functionality"`
+	Scope             types.Int64  `tfsdk:"scope"`
+	RequiresElevation types.Bool   `tfsdk:"requires_elevation"`
+}
+
+func (s inputRole) getListOfScopedFunctionalities() []eva.RoleScopedFunctionality {
+	var scopedFunctionalities []eva.RoleScopedFunctionality
+
+	for _, scopedFunctionality := range s.ScopedFunctionalities {
+		scopedFunctionalities = append(scopedFunctionalities, eva.RoleScopedFunctionality{
+			Functionality:     scopedFunctionality.Functionality.Value,
+			Scope:             scopedFunctionality.Scope.Value,
+			RequiresElevation: scopedFunctionality.RequiresElevation.Value,
+		})
+	}
+
+	return scopedFunctionalities
+}
+
+func (s inputRole) setListOfScopedFunctionalities(scopedFunctionalities []eva.RoleScopedFunctionality) {
+	for _, scopedFunctionality := range scopedFunctionalities {
+		s.ScopedFunctionalities = append(s.ScopedFunctionalities, inputRoleScopedFunctionality{
+			Functionality:     types.String{Value: scopedFunctionality.Functionality},
+			Scope:             types.Int64{Value: scopedFunctionality.Scope},
+			RequiresElevation: types.Bool{Value: scopedFunctionality.RequiresElevation},
+		})
+	}
 }
 
 func (r roleResource) Create(ctx context.Context, req tfsdk.CreateResourceRequest, resp *tfsdk.CreateResourceResponse) {
-	var data inputRoleData
+	var data inputRole
 
 	diags := req.Config.Get(ctx, &data)
 	resp.Diagnostics.Append(diags...)
@@ -76,18 +131,28 @@ func (r roleResource) Create(ctx context.Context, req tfsdk.CreateResourceReques
 		return
 	}
 
-	client_resp, err := r.provider.evaClient.CreateRole(ctx, eva.CreateRoleRequest{
+	createdRole, createRoleErr := r.provider.evaClient.CreateRole(ctx, eva.CreateRoleRequest{
 		Name:     data.Name.Value,
 		UserType: data.UserType.Value,
 		Code:     data.Code.Value,
 	})
 
-	if err != nil {
-		resp.Diagnostics.AddError("Creating role unit failed.", fmt.Sprintf("Unable to create role, got error: %s", err))
+	if createRoleErr != nil {
+		resp.Diagnostics.AddError("Creating role unit failed.", fmt.Sprintf("Unable to create role, got error: %s", createRoleErr))
 		return
 	}
 
-	data.ID = types.Int64{Value: client_resp.ID}
+	data.ID = types.Int64{Value: createdRole.ID}
+
+	_, attachPermissionsToRoleErr := r.provider.evaClient.AttachFunctionalitiesToRole(ctx, eva.AttachFunctionalitiesToRoleRequest{
+		RoleID:                data.ID.Value,
+		ScopedFunctionalities: data.getListOfScopedFunctionalities(),
+	})
+
+	if attachPermissionsToRoleErr != nil {
+		resp.Diagnostics.AddError("Creating role permissions failed.", fmt.Sprintf("Unable to create role permisions, got error: %s", attachPermissionsToRoleErr))
+		return
+	}
 
 	tflog.Trace(ctx, "Created a new role.")
 
@@ -96,7 +161,7 @@ func (r roleResource) Create(ctx context.Context, req tfsdk.CreateResourceReques
 }
 
 func (r roleResource) Read(ctx context.Context, req tfsdk.ReadResourceRequest, resp *tfsdk.ReadResourceResponse) {
-	var data inputRoleData
+	var data inputRole
 
 	diags := req.State.Get(ctx, &data)
 	resp.Diagnostics.Append(diags...)
@@ -105,7 +170,7 @@ func (r roleResource) Read(ctx context.Context, req tfsdk.ReadResourceRequest, r
 		return
 	}
 
-	client_resp, err := r.provider.evaClient.GetRole(ctx, eva.GetRoleRequest{
+	roleData, err := r.provider.evaClient.GetRole(ctx, eva.GetRoleRequest{
 		ID: data.ID.Value,
 	})
 
@@ -114,16 +179,17 @@ func (r roleResource) Read(ctx context.Context, req tfsdk.ReadResourceRequest, r
 		return
 	}
 
-	data.Name = types.String{Value: client_resp.Result.Name}
-	data.UserType = types.Int64{Value: client_resp.Result.UserType}
-	data.Code = types.String{Value: client_resp.Result.Code}
+	data.Name = types.String{Value: roleData.Result.Name}
+	data.UserType = types.Int64{Value: roleData.Result.UserType}
+	data.Code = types.String{Value: roleData.Result.Code}
+	data.setListOfScopedFunctionalities(roleData.Result.ScopedFunctionalities)
 
 	diags = resp.State.Set(ctx, &data)
 	resp.Diagnostics.Append(diags...)
 }
 
 func (r roleResource) Update(ctx context.Context, req tfsdk.UpdateResourceRequest, resp *tfsdk.UpdateResourceResponse) {
-	var data inputRoleData
+	var data inputRole
 
 	diags := req.Plan.Get(ctx, &data)
 	resp.Diagnostics.Append(diags...)
@@ -144,12 +210,41 @@ func (r roleResource) Update(ctx context.Context, req tfsdk.UpdateResourceReques
 		return
 	}
 
+	roleData, getRoleErr := r.provider.evaClient.GetRole(ctx, eva.GetRoleRequest{
+		ID: data.ID.Value,
+	})
+
+	if getRoleErr != nil {
+		resp.Diagnostics.AddError("Updating role permissions failed.", fmt.Sprintf("Unable to get role, got error: %s", getRoleErr))
+		return
+	}
+
+	_, detachErr := r.provider.evaClient.DetachFunctionalitiesFromRole(ctx, eva.DetachFunctionalitiesFromRoleRequest{
+		RoleID:                data.ID.Value,
+		ScopedFunctionalities: roleData.Result.ScopedFunctionalities,
+	})
+
+	if detachErr != nil {
+		resp.Diagnostics.AddError("Updating role permissions failed.", fmt.Sprintf("Unable to detach current role permissions, got error: %s", detachErr))
+		return
+	}
+
+	_, attachErr := r.provider.evaClient.AttachFunctionalitiesToRole(ctx, eva.AttachFunctionalitiesToRoleRequest{
+		RoleID:                data.ID.Value,
+		ScopedFunctionalities: data.getListOfScopedFunctionalities(),
+	})
+
+	if attachErr != nil {
+		resp.Diagnostics.AddError("Updating role permissions failed.", fmt.Sprintf("Unable to attach new role permissions, got error: %s", detachErr))
+		return
+	}
+
 	diags = resp.State.Set(ctx, &data)
 	resp.Diagnostics.Append(diags...)
 }
 
 func (r roleResource) Delete(ctx context.Context, req tfsdk.DeleteResourceRequest, resp *tfsdk.DeleteResourceResponse) {
-	var data inputRoleData
+	var data inputRole
 
 	diags := req.State.Get(ctx, &data)
 	resp.Diagnostics.Append(diags...)
